requirements-test.txt
jinja2>=2.11
coverage==4.4.1
mock>=1.0.1
flake8>=2.1.0
tox>=1.7.0
codecov>=2.0.0
pytest-cov>=2.8.1
